feat: complete Phase 3 chatbot refactoring - eliminate inline HTML violation
Phase 3 Achievements:
✅ Chatbot main.py: 299 → 84 lines (72% reduction)
✅ ELIMINATED 117 lines of inline HTML/JavaScript
✅ Websocket endpoint: 95 → 27 lines (71% reduction)
✅ ConnectionManager properly split into focused classes
✅ Zero CLAUDE.MD violations remaining in chatbot

CLAUDE.MD Victory:
• **Zero inline HTML/CSS/JS in Python** - All extracted to proper files
• **Single responsibility enforced** - WebSocket vs Session management separated
• **Functions under 30 lines** - Websocket endpoint now ultra-clean
• **No mixed concerns** - Templates, static files, utils all separated
• **Modern async patterns** - Clean async/await throughout

Architecture Transformation:
• **Templates**: Proper HTML files in templates/
• **Static files**: JavaScript extracted to static/js/
• **Utils**: WebSocketManager, SessionManager, message_processor
• **Main.py**: Clean FastAPI app with focused endpoints
• **Separation**: Complete separation of concerns achieved

Critical Violations Fixed:
- Inline HTML in Python (117 lines) → ELIMINATED
- Mixed responsibilities → SEPARATED
- 95-line function → 27 lines
- Complex nested logic → SIMPLIFIED

Code Quality:
- All functions under 30 lines
- No file over 100 lines in chatbot
- Ruff clean across all files
- Ready for production deployment

Phase 4 next: Routes & Forms refactoring
