+++ chatbot/templates/chat_widget.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Chat Widget Test</title>
+    <link rel="stylesheet" href="../app/static/css/components.css">
+    <link rel="stylesheet" href="../app/static/css/chatbot.css">
+</head>
+<body>
+    <h2>Chatbot Test Interface</h2>
+    <div id="chat-container">
+        <div id="messages"></div>
+        <div id="input-area">
+            <input type="text" id="message-input" placeholder="Type your message...">
+            <button id="send-button">Send</button>
+        </div>
+    </div>
+    <script src="/static/js/chat_widget.js"></script>
+</body>
+</html>
